+++ src/Neo.VM/Types/ByteString.cs
@@ -9,7 +9,6 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
-using Neo.Extensions;
 using System;
 using System.Diagnostics;
 using System.Numerics;
@@ -90,7 +89,7 @@ namespace Neo.VM.Types
 
             fixed (byte* k = bytes)
             {
-                return (int)UnsafeData.HashBytes(k, bytes.Length);
+                return (int)Unsafe.HashBytes(k, bytes.Length);
             }
         }
